[SC-166] Add filter settings view
